keys: Move SequenceCacheKey decoding to keys package
Fixes #4547.

The `decodeSequenceCacheKey` is more at home in the `keys` package
than the `storage` package.
